Display list of contacts and click on contact to manage them
