+++ guilds/video/ui/family/index.tsx
@@ -1,41 +1,81 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { Routes, Route, Link } from 'react-router-dom';
 import { Card } from '@openconnect/ui';
-import Contacts from './manage_contacts';
+import ManageContact from './manage_contact';
+
+interface IncarceratedPerson {
+  id: string;
+  firstName: string;
+  lastName: string;
+  facilityId?: string;
+}
+
+interface ApprovedContactItem {
+  id: string;
+  incarceratedPerson: IncarceratedPerson;
+  relationship?: string;
+}
 
 function VideoHome() {
+  const [contacts, setContacts] = useState<ApprovedContactItem[] | null>(null);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect(() => {
+    let mounted = true;
+    async function load() {
+      setLoading(true);
+      try {
+        // @ts-ignore
+        const token: string | null = typeof localStorage !== 'undefined' ? localStorage.getItem('token') : null;
+        const res = await fetch('/api/video/approved-contacts', {
+          headers: token ? { Authorization: `Bearer ${token}` } : {},
+        });
+        if (!res.ok) throw new Error(`HTTP ${res.status} ${res.statusText}`);
+        const json: any = await res.json();
+        const data: ApprovedContactItem[] = json.data || json;
+        if (mounted) setContacts(data);
+      } catch (err: any) {
+        if (mounted) setError(err.message || 'Failed to load contacts');
+      } finally {
+        if (mounted) setLoading(false);
+      }
+    }
+    load();
+    return () => { mounted = false; };
+  }, []);
+
   return (
     <div className="space-y-4">
       <h1 className="text-2xl font-bold text-gray-900">Video Visits</h1>
+
       <Card padding="lg">
-        <div className="text-center py-8">
-          <span className="text-6xl mb-4 block">📹</span>
-          <h2 className="text-xl font-semibold mb-2">Schedule & Join Video Visits</h2>
-          <div className="mt-6 flex flex-col items-center gap-3">
-            <Link
-              to="schedule"
-              className="inline-flex items-center px-4 py-2 bg-blue-600 text-white rounded-md shadow-sm hover:bg-blue-700"
-            >
-              Schedule video call
-            </Link>
-            <Link
-              to="calls"
-              className="inline-flex items-center px-4 py-2 border border-gray-300 rounded-md text-gray-700 hover:bg-gray-50"
-            >
-              View scheduled calls
-            </Link>
-            <Link
-              to="past-calls"
-              className="inline-flex items-center px-4 py-2 border border-gray-300 rounded-md text-gray-700 hover:bg-gray-50"
-            >
-              View past calls
-            </Link>
-            <Link
-              to="manage"
-              className="inline-flex items-center px-4 py-2 border border-gray-300 rounded-md text-gray-700 hover:bg-gray-50"
-            >
-              Manage contacts
-            </Link>
+        <div className="space-y-4">
+          <p className="text-gray-600">Select a contact to schedule or manage video visits.</p>
+
+          {loading && <div className="text-gray-500">Loading...</div>}
+          {error && <div className="text-red-600">{error}</div>}
+
+          {!loading && contacts && contacts.length === 0 && (
+            <div className="text-gray-500">No approved contacts found.</div>
+          )}
+
+          <div className="grid gap-3">
+            {contacts && contacts.map((c) => (
+              <Link
+                key={c.id}
+                to={`manage_contact/${c.incarceratedPerson.id}`}
+                className="flex items-center justify-between p-4 border rounded-md hover:bg-gray-50 transition-colors"
+              >
+                <div>
+                  <div className="font-medium text-gray-900">
+                    {c.incarceratedPerson.firstName} {c.incarceratedPerson.lastName}
+                  </div>
+                  {c.relationship && <div className="text-sm text-gray-500">{c.relationship}</div>}
+                </div>
+                <div className="text-gray-400">→</div>
+              </Link>
+            ))}
           </div>
         </div>
       </Card>
@@ -47,7 +87,7 @@ export default function VideoFamily() {
   return (
     <Routes>
       <Route index element={<VideoHome />} />
-      <Route path="manage" element={<Contacts />} />
+      <Route path="manage_contact/:contactId" element={<ManageContact />} />
     </Routes>
   );
 }
